requirements.txt
streamlit
langchain-google-genai
langchain==0.3.7
langchain-community==0.3.7
langchain-text-splitters
chromadb
pypdf
python-dotenv
bcrypt
sentence-transformers
